bug: fix contents_first with root symlink
This fixes two incorrect behaviors when `contents_first` is on and the
root path is a symlink:

  1. The root path was returned first in the iterator instead of last.
  2. Some subdirectories were returned out of order.

The issue was that root symlinks were returned immediately rather than
being pushed onto the `deferred_dirs` vec. That lead to `deferred_dirs`
and `depth` being out of sync, which lead to deferred directories being
processed one ascent too late.

TO DO:
  [ ] Remove FIXME
  [ ] Consider commenting changed code; it seems somewhat hard to follow
      to my sleep deprived brain.

Fixes: #163