bug: fix contents_first with same_file_system
This fixes a bug that occurred when both the `contents_first` and
`same_file_system` options were on and a subdirectory was on another
filesystem. Such a directory would be added to the `deferred_dirs` vec,
but would not actually be descended into, leading to it being emitted at
the wrong time.

TO DO:
  [] Actual test